Castin position (int)value

+++ NationalLibrary/Data/Location.cs
@@ -34,7 +34,7 @@ namespace NationalLibrary.Data
 		public string Room { get => room; set { room = DataController.CheckStrings(value, "stanza"); } }
 		public string Schaffold { get => scaffhold; set { scaffhold = DataController.CheckScaffholdOrShelf(value, "scaffale"); } }
 		public string Shelf { get => shelf; set { shelf = DataController.CheckScaffholdOrShelf(value, "ripiano"); } }
-		public int? Position { get => position; set { position = value; } }
+		public int? Position { get => position; set { position = (int)value; } }
 		#endregion
 
 		// Relation Location 1-1 Book(FK)
